# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sqordia API",
    "description": "Comprehensive API for business plan generation and management platform with Google OAuth authentication support",
    "version": "1.0.0",
    "contact": {
      "name": "Sqordia API Support",
      "email": "support@sqordia.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.sqordia.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/organizations/{organizationId}": {
      "get": {
        "tags": ["Organizations"],
        "summary": "Get organization details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/business-plans/{businessPlanId}": {
      "get": {
        "tags": ["Business Plans"],
        "summary": "Get business plan details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "businessPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business plan details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessPlanDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/obnl/plans/{planId}": {
      "get": {
        "tags": ["OBNL Business Plans"],
        "summary": "Get OBNL plan details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OBNL plan details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OBNLPlanDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BusinessPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "businessType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "OBNLPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "mission": {
            "type": "string"
          },
          "vision": {
            "type": "string"
          },
          "obnlType": {
            "type": "string"
          },
          "targetAudience": {
            "type": "string"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fundingSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "complianceRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "complianceStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}